split api serializers
