+++ netbox_access_lists/api/views.py
@@ -8,8 +8,13 @@ from django.db.models import Count
 from netbox.api.viewsets import NetBoxModelViewSet
 
 from .. import filtersets, models
-from .serializers import (AccessListSerializer, ACLExtendedRuleSerializer,
-                          ACLStandardRuleSerializer)
+from .serializers import AccessListSerializer, ACLStandardRuleSerializer, ACLExtendedRuleSerializer
+
+__all__ = [
+    'AccessListViewSet',
+    'ACLStandardRuleViewSet',
+    'ACLExtendedRuleViewSet',
+]
 
 
 class AccessListViewSet(NetBoxModelViewSet):
